Seperate post deletion into its own component
