+++ devblog/devblog/ClientApp/src/components/Post.tsx
@@ -1,19 +1,13 @@
 import IPost from "../interfaces/IPost";
 import Comment from "./Comment";
 import CreateComment from "./CreateComment";
+import DeletePost from "./DeletePost";
 import "./Post.css";
 
 const Post = (props: IPost) => {
-    const handleDelete = () => {
-        fetch(`api/posts/${props.id}`, {
-            method: "DELETE",
-            headers: { "Content-Type": "application/json" }
-        });
-    };
-
     return (
         <div className="post">
-            <button onClick={handleDelete}>Delete</button>
+            <DeletePost id={props.id} />
             <div className="post-info">
                 <span>{props.updateNum}</span>
                 <span>{props.date}</span>
